+++ src/main/java/search/SearchMechanism.java
@@ -1,0 +1,5 @@
+package search;
+
+public interface SearchMechanism {
+    void findPeople();
+}
