+++ src/main/java/utility/UserPromptUtility.java
@@ -1,0 +1,5 @@
+package utility;
+
+public interface UserPromptUtility {
+    void startMenu ();
+}
